+++ src/inc.h
@@ -73,7 +73,7 @@ struct _IncSession
 	GHashTable *folder_table;	/* table of destination folders */
 	GHashTable *tmp_folder_table;	/* for progressive update */
 
-	gint cur_total_bytes;
+	gint64 cur_total_bytes;
 
 	gpointer data;
 };
